BGDIINF_SB-1490 moved and added unit tests
* moved CollectionsModelTemporalExtentTestCase to test_collection_extent.py
* renamed test_item_to_collection_propagation.py to test_collections_extent.py
* added unittests for creation of collections in test_collections_model.py
